requirements.txt
python-dotenv~=1.0.0
Flask==2.2.2
Flask-Cors==3.0.10
keras==2.8.0
Keras-Preprocessing==1.1.2
numpy==1.23.5
opencv-python==4.6.0.66
pandas==1.4.3
requests==2.28.1
scikit-image==0.19.3
Pillow~=9.2.0
Werkzeug~=2.2.2
tensorflow~=2.8.0
librosa==0.8.1
matplotlib~=3.5.3